+ scalar
